Moved tests from ThemeValidationUtilTest to TaskRunDevBundleBuildTest (thanks Mikhail for catching it)
